rwork structure to make adding new options easier
Before, operators were typed by every component type, leading to a lot
of copy pasting of type names, and a lot of work to add a new field as
everything needed to change. Now we use unpacking to efficient replace
options for use in an operator copy, and use a handy type utility to
express how the type is changing.