make tests use ts-expect-error
To test edge cases that happened in javascript but that woudl be caught
in typescript, tests used ugly casts to fool the compiler. This diff
makes those tests use @ts-expect-error instead to be ensure that we
normally error in that case, and then the tests also expect that a
runtime error is thrown.