simplify typings
In the past this library used complex strings of `Parameters<...>` to
extract the typings over operators to do the intersection when an
operator took multiple subfunctions that would all influence a type.

This updates those to use simple `infer` clauses making the types much
easier to read.

This also adds default types to the various operators that are the
typing returned by vanilla constructors, although as of right now, those
aren't exposed